+++ tests/test_config.py
@@ -1,5 +1,6 @@
 from stac_check.lint import Linter
 
+
 def test_linter_config_file():
     file = "sample_files/1.0.0/core-item.json"
     linter = Linter(file)
@@ -8,7 +9,7 @@ def test_linter_config_file():
     assert linter.config["linting"]["searchable_identifiers"] == True
     assert linter.create_best_practices_dict()["searchable_identifiers"] == [
         f"Item name '{linter.object_id}' should only contain Searchable identifiers",
-        "Identifiers should consist of only lowercase characters, numbers, '_', and '-'"
+        "Identifiers should consist of only lowercase characters, numbers, '_', and '-'",
     ]
 
     # Load config file
@@ -16,6 +17,7 @@ def test_linter_config_file():
 
     assert linter.config["linting"]["searchable_identifiers"] == False
     assert "searchable_identifiers" not in linter.create_best_practices_dict()
+
 
 def test_linter_max_links():
     file = "sample_files/1.0.0/core-item-bloated.json"
@@ -27,6 +29,3 @@ def test_linter_max_links():
     # Load config file
     linter = Linter(file, config_file="tests/test.config.yml")
     assert "bloated_links" not in linter.create_best_practices_dict()
-
-
-